+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [[Unreleased]]
 ### Fixed
 - Improved documentation on using websocket clients.
+- Fixes an issue sending NoRippleCheck requests
 
 ## [1.1.1] - 2021-07-02
 ### Fixed